First functional build on tauri v2.
