+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "headphones-toolbox",
   "private": true,
-  "version": "0.0.6",
+  "version": "0.0.7",
   "type": "module",
   "scripts": {
     "dev": "vite",
@@ -12,7 +12,11 @@
   "dependencies": {
     "@quasar/extras": "^1.16.3",
     "@quasar/vite-plugin": "^1.3.1",
-    "@tauri-apps/api": "^1.4.0",
+    "@tauri-apps/api": "^2.0.0-beta.11",
+    "@tauri-apps/plugin-dialog": "^2.0.0-beta.3",
+    "@tauri-apps/plugin-fs": "^2.0.0-beta.3",
+    "@tauri-apps/plugin-os": "^2.0.0-beta.3",
+    "@tauri-apps/plugin-shell": "^2.0.0-beta.4",
     "@vitejs/plugin-vue": "^4.2.1",
     "chart.js": "^4.3.0",
     "lodash.debounce": "^4.0.8",
@@ -23,7 +27,7 @@
     "vue-chartjs": "^5.2.0"
   },
   "devDependencies": {
-    "@tauri-apps/cli": "^1.4.0",
+    "@tauri-apps/cli": "^2.0.0-beta.17",
     "typescript": "^4.8.2",
     "vite": "^4.2.1"
   }
