+++ src-tauri/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "headphones-toolbox"
-version = "0.0.5"
+version = "0.0.7"
 description = "An application configure Ploopy Headphones."
 authors = ["George Norton"]
 license = ""
@@ -10,10 +10,10 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [build-dependencies]
-tauri-build = { version = "1.3", features = [] }
+tauri-build = { version = "2.0.0-beta", features = [] }
 
 [dependencies]
-tauri = { version = "1.3", features = [ "os-all", "dialog-all", "fs-all", "path-all", "shell-open", "window-close", "window-hide", "window-maximize", "window-minimize", "window-show", "window-start-dragging", "window-unmaximize", "window-unminimize"] }
+tauri = { version = "2.0.0-beta", features = [] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 rusb = "0.9.2"
@@ -22,8 +22,16 @@ window-shadows = "0.2.1"
 simplelog = "0.12.1"
 log = "0.4.20"
 parking_lot = "0.12.1"
+tauri-plugin-fs = "2.0.0-beta.7"
+tauri-plugin-shell = "2.0.0-beta.5"
+tauri-plugin-dialog = "2.0.0-beta.7"
+tauri-plugin-os = "2.0.0-beta.4"
 
 [features]
 # this feature is used for production builds or when `devPath` points to the filesystem
 # DO NOT REMOVE!!
 custom-protocol = ["tauri/custom-protocol"]
+
+[lib]
+name = "app_lib"
+crate-type = ["staticlib", "cdylib", "rlib"]
