+++ src-tauri/tauri.conf.json
@@ -2,73 +2,31 @@
   "build": {
     "beforeDevCommand": "yarn dev",
     "beforeBuildCommand": "yarn build",
-    "devPath": "http://localhost:1420",
-    "distDir": "../dist",
-    "withGlobalTauri": true
+    "frontendDist": "../dist",
+    "devUrl": "http://localhost:1420"
   },
-  "package": {
-    "productName": "Ploopy Headphones Toolbox",
-    "version": "0.0.6"
+  "bundle": {
+    "active": true,
+    "icon": [
+      "icons/32x32.png",
+      "icons/128x128.png",
+      "icons/128x128@2x.png",
+      "icons/icon.icns",
+      "icons/icon.ico"
+    ],
+    "targets": "all",
+    "resources": [
+      "resources/*"
+    ]
   },
-  "tauri": {
-    "allowlist": {
-      "all": false,
-      "shell": {
-        "all": false,
-        "open": ".*log"
-      },
-      "dialog": {
-        "all": true,
-        "ask": true,
-        "confirm": true,
-        "message": true,
-        "open": true,
-        "save": true
-      },
-      "window": {
-        "all": false,
-        "close": true,
-        "hide": true,
-        "show": true,
-        "maximize": true,
-        "minimize": true,
-        "unmaximize": true,
-        "unminimize": true,
-        "startDragging": true
-      },
-      "fs": {
-        "all": true,
-        "scope": ["$APPDATA/", "$APPDATA/**", "$RESOURCE/resources/*", "**", "$APPLOGDIR/", "$APPLOGDIR/**"]
-      },
-      "os": {
-        "all": true 
-      },
-      "path": {
-        "all": true
-      }
-    },
-    "bundle": {
-      "active": true,
-      "icon": [
-        "icons/32x32.png",
-        "icons/128x128.png",
-        "icons/128x128@2x.png",
-        "icons/icon.icns",
-        "icons/icon.ico"
-      ],
-      "identifier": "app.ploopy.headphones",
-      "resources": [
-        "resources/*"
-      ],
-      
-      "targets": "all"
-    },
-    "security": {
-      "csp": "default-src 'self'"
-    },
-    "updater": {
-      "active": false
-    },
+  "productName": "Ploopy Headphones Toolbox",
+  "version": "0.0.7",
+  "identifier": "app.ploopy.headphones",
+  "plugins": {
+    "fs": {}
+  },
+  "app": {
+    "withGlobalTauri": true,
     "windows": [
       {
         "fullscreen": false,
@@ -78,6 +36,9 @@
         "height": 720,
         "decorations": false
       }
-    ]
+    ],
+    "security": {
+      "csp": "default-src 'self'; connect-src ipc: http://ipc.localhost"
+    }
   }
 }
